Google Maps now starts at the location specified by the latitude and longitude stored in the database. However, there is no way to update the database or Lat/Lng TextViews because the app requirements specifically say not to (which I find a little strange but whatever).
Also, latitude and longitude now use the 'double' datatype because I realised how dumb it is to save them as an integer datatype like 'long'.

Unless I think of something small, it's looking like this is the last commit for the location requirements. Sweet.

+++ app/src/main/java/database/ReceiptCursorWrapper.java
@@ -19,8 +19,8 @@ public class ReceiptCursorWrapper extends CursorWrapper {
         String shop = getString(getColumnIndex(ReceiptDbSchema.ReceiptTable.Cols.SHOP)); //edit
         String comment = getString(getColumnIndex(ReceiptDbSchema.ReceiptTable.Cols.COMMENT)); //edit2
         long date = getLong(getColumnIndex(ReceiptDbSchema.ReceiptTable.Cols.DATE));
-        long latitude = getLong(getColumnIndex(ReceiptDbSchema.ReceiptTable.Cols.LATITUDE));
-        long longitude = getLong(getColumnIndex(ReceiptDbSchema.ReceiptTable.Cols.LONGITUDE));
+        double latitude = getLong(getColumnIndex(ReceiptDbSchema.ReceiptTable.Cols.LATITUDE));
+        double longitude = getLong(getColumnIndex(ReceiptDbSchema.ReceiptTable.Cols.LONGITUDE));
 
         Receipt receipt = new Receipt(UUID.fromString(uuidString));
         receipt.setTitle(title);
